requirements.txt
six~=1.16.0
numpy~=2.1.0
llama-cpp-python~=0.2.89
llama-cpp-python[server]
torch==2.4.0

